input/requirements.txt
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.1
PyMySQL==1.1.0
pydantic==2.6.4
dbutils==3.1.0
